+++ tests/TestCase.php
@@ -53,6 +53,11 @@ class TestCase extends \PHPUnit_Framework_TestCase
             'id' => 'testapp',
             'basePath' => __DIR__,
             'vendorPath' => $this->getVendorPath(),
+            'components' => [
+                'urlManager' => [
+                    'scriptUrl' => '/index.php'
+                ],
+            ],
         ], $config));
     }